requirements/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.1.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
huggingface-hub>=0.20.0,<1.0
transformers>=4.30.0
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
markitdown[pdf]>=0.0.1